requirements.txt
google-generativeai
streamlit
python-dotenv
streamlit-chat